# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LRMC — Legacy Rental Management Consortium API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers:
    get:
      tags: [ususu]
      operationId: listDrivers
      summary: List Ususu drivers
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      tags: [ususu]
      operationId: createDriver
      summary: Apply to become an Ususu driver (free membership)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    patch:
      tags: [ususu]
      operationId: updateDriver
      summary: Update a driver's status
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /rides:
    get:
      tags: [ususu]
      operationId: listRides
      summary: List rides
      parameters:
        - name: riderId
          in: query
          required: false
          schema:
            type: string
        - name: driverId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ride"
    post:
      tags: [ususu]
      operationId: createRide
      summary: Request a ride
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
  /rides/{id}:
    patch:
      tags: [ususu]
      operationId: updateRide
      summary: Update ride status (accept/complete/cancel) — settles fare into LRMC treasury
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"

  # ---------------- Payments ----------------
...
  /youth-employment-records:
    get:
      tags: [youth]
      operationId: listYouthEmploymentRecords
      summary: List youth employment records
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YouthEmploymentRecord"
    post:
      tags: [youth]
      operationId: createYouthEmploymentRecord
      summary: Apply to the youth employment initiative
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YouthEmploymentRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YouthEmploymentRecord"
  /youth-employment-records/{id}:
    patch:
      tags: [youth]
      operationId: updateYouthEmploymentRecord
      summary: Update a youth employment record's status
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YouthEmploymentRecordUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YouthEmploymentRecord"

  # ---------------- Prospecting ----------------
...
  /settlement-obligations:
    get:
      tags: [settlement]
      operationId: listSettlementObligations
      summary: List settlement obligations (driver/contractor/vendor payouts)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SettlementObligation"
    post:
      tags: [settlement]
      operationId: createSettlementObligation
      summary: Schedule a settlement obligation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementObligationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementObligation"
  /settlement-obligations/{id}:
    patch:
      tags: [settlement]
      operationId: updateSettlementObligation
      summary: Mark a settlement obligation paid/overdue/cancelled
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementObligationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementObligation"

  # ---------------- Dashboards ----------------
...
components:
  parameters:
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Driver:
      type: object
      required: [id, userId, vehicleInfo, licenseNumber, status, rating, createdAt]
      properties:
        id: { type: integer }
        userId: { type: string }
        vehicleInfo: { type: string }
        licenseNumber: { type: string }
        status: { type: string }
        rating: { type: number }
        createdAt: { type: string, format: date-time }
    DriverInput:
      type: object
      required: [userId]
      properties:
        userId: { type: string }
        vehicleInfo: { type: string }
        licenseNumber: { type: string }
    DriverUpdate:
      type: object
      properties:
        status: { type: string }
        vehicleInfo: { type: string }
        licenseNumber: { type: string }
        rating: { type: number }
    Ride:
      type: object
      required: [id, riderId, driverId, pickup, dropoff, fareGmd, status, requestedAt, completedAt]
      properties:
        id: { type: integer }
        riderId: { type: string }
        driverId: { type: [integer, 'null'] }
        pickup: { type: string }
        dropoff: { type: string }
        fareGmd: { type: integer }
        status: { type: string }
        requestedAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
    RideInput:
      type: object
      required: [riderId, pickup, dropoff, fareGmd]
      properties:
        riderId: { type: string }
        pickup: { type: string, minLength: 1 }
        dropoff: { type: string, minLength: 1 }
        fareGmd: { type: integer }
    RideUpdate:
      type: object
      properties:
        driverId: { type: integer }
        status: { type: string }
        completedAt: { type: [string, 'null'], format: date-time }

    # ---- Payments ----
...
    YouthEmploymentRecord:
      type: object
      required: [id, fullName, email, phone, program, status, placementCompany, createdAt]
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: [string, 'null'] }
        phone: { type: [string, 'null'] }
        program: { type: string }
        status: { type: string }
        placementCompany: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    YouthEmploymentRecordInput:
      type: object
      required: [fullName]
      properties:
        fullName: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        program: { type: string }
    YouthEmploymentRecordUpdate:
      type: object
      properties:
        status: { type: string }
        placementCompany: { type: string }

    # ---- Prospecting ----
...
    SettlementObligation:
      type: object
      required: [id, payeeType, payeeName, amount, currency, dueDate, status, createdAt]
      properties:
        id: { type: integer }
        payeeType: { type: string }
        payeeName: { type: string }
        amount: { type: number }
        currency: { type: string }
        dueDate: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    SettlementObligationInput:
      type: object
      required: [payeeType, payeeName, amount, dueDate]
      properties:
        payeeType: { type: string }
        payeeName: { type: string, minLength: 1 }
        amount: { type: number }
        currency: { type: string }
        dueDate: { type: string }
    SettlementObligationUpdate:
      type: object
      properties:
        status: { type: string }
        amount: { type: number }
        dueDate: { type: string }

    # ---- Dashboards ----
...